+++ src/App.jsx
@@ -1,12 +1,14 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 
 import TodoList from './features/TodoList/TodoList.component';
 import TodoForm from './features/TodoForm/TodoForm.component';
 
 function App() {
   const [todoList, setTodoList] = useState([]);
-  const [errorMessage, setErrorMessage] = useState('')
+  const [errorMessage, setErrorMessage] = useState('');
   const [isLoading, setIsLoading] = useState(false);
+  const url = `https://api.airtable.com/v0/${import.meta.env.VITE_BASE_ID}/${import.meta.env.VITE_TABLE_NAME}`;
+  const token = `Bearer ${import.meta.env.VITE_PAT}`;
 
   const addTodo = title => {
     const newTodo = {
@@ -33,6 +35,40 @@ function App() {
     });
     setTodoList(updatedTodoList);
   };
+
+  useEffect(() => {
+    const fetchTodos = async () => {
+      setIsLoading(true);
+      const options = {
+        method: 'GET',
+        headers: { Authorization: token },
+      };
+      try {
+        const resp = await fetch(url, options);
+        if (!resp.ok) throw new Error(resp.message);
+
+        const { records } = await resp.json();
+        const todos = records.map(record => {
+          const todo = {
+            id: record.id,
+            ...record.fields,
+          };
+          if (!todo.isCompleted) {
+            todo.isCompleted = false;
+          }
+          console.log(todo);
+          return todo;
+        });
+        setTodoList([...todos]);
+      } catch (err) {
+        setErrorMessage(err.message);
+      }
+       finally {
+        setIsLoading(false);
+       }
+    };
+    fetchTodos();
+  }, []);
   return (
     <div>
       <h1 className="todos-title">My Todos</h1>